kernel: qca-ssdk: update to 12.4.r1
Update SSDK to top of 12.4.r1 upstream, it introduces support for building
SSDK as a proper out of tree module instead of the current hacky way,basic
kernel 6.1 support is included as well.

It allows dropping most of the patches since they have either been fixed
upstream to work with 5.15 or by building as an out of tree module we are
not hitting them anymore, it also fixes building with make 4.4 and newer.

Signed-off-by: Robert Marko <robimarko@gmail.com>