ssdk: switch back to NHSS.QSDK.12.4
NOTE: Incorrectly stated version 3.0 which is the SSDK version, and not
'NHSS.QSDK' version with is an entire release.

SSDK version in NHSS.QSDK.12.4.5.r3 is mostly for newer switches and doesn't bring anything new
to existing switches. Switch back to version NHSS.QSDK.12.4 to stay in sync with
upstream.