qualcommax: consolidate symbols, revert built-in
Symbols for crypto, ktls, netfilter are removed from from being compiled
into the kernel and instead left up to the user to compile as modules,
as is the case upstream. Other unused qca features were also removed as
they don't apply to ipq807x/6018 and are not upstream.

Common symbols between 6.1/6.6 that are required for NSS related
features have been consolidated into 'config-default'.

There is an average 300kb of savings in the final kernel image as well
decreased compile time.